+++ tests/test_models.py
@@ -339,11 +339,9 @@ class TestModelOperators:
         a = self.mod_hti_a + self.mod_hti_a
         assert_allclose(a.property_x, self.mod_hti_b.property_x)
         assert_allclose(a.property_y, self.mod_hti_b.property_y)
-        # assert a.property_z.base is a.property_x.base
 
         a = self.mod_vti_a + self.mod_vti_a
         assert_allclose(a.property_x, self.mod_vti_b.property_x)
-        # assert a.property_y.base is a.property_x.base
         assert_allclose(a.property_z, self.mod_vti_b.property_z)
 
         a = self.mod_tri_a + self.mod_tri_a
@@ -366,11 +364,9 @@ class TestModelOperators:
         a = self.mod_hti_b - self.mod_hti_a
         assert_allclose(a.property_x, self.mod_hti_a.property_x)
         assert_allclose(a.property_y, self.mod_hti_a.property_y)
-        # assert a.property_z.base is a.property_x.base
 
         a = self.mod_vti_b - self.mod_vti_a
         assert_allclose(a.property_x, self.mod_vti_a.property_x)
-        # assert a.property_y.base is a.property_x.base
         assert_allclose(a.property_z, self.mod_vti_a.property_z)
 
         a = self.mod_tri_b - self.mod_tri_a
@@ -411,13 +407,15 @@ class TestModelOperators:
         assert model_new3 == self.mod_mu_a
         assert model_new4 == self.mod_epsilon_a
 
-        assert (model_new1.property_x.base is not
-                self.mod_shape.property_x.base)
-        assert model_new2.property_y.base is not self.mod_tri_a.property_y.base
-        assert model_new2.property_z.base is not self.mod_tri_a.property_z.base
-        assert model_new3.mu_r.base is not self.mod_mu_a.mu_r.base
-        assert (model_new4.epsilon_r.base is not
-                self.mod_epsilon_a.epsilon_r.base)
+        assert not np.may_share_memory(model_new1.property_x,
+                                       self.mod_shape.property_x)
+        assert not np.may_share_memory(model_new2.property_y,
+                                       self.mod_tri_a.property_y)
+        assert not np.may_share_memory(model_new2.property_z,
+                                       self.mod_tri_a.property_z)
+        assert not np.may_share_memory(model_new3.mu_r, self.mod_mu_a.mu_r)
+        assert not np.may_share_memory(model_new4.epsilon_r,
+                                       self.mod_epsilon_a.epsilon_r)
 
     def test_dict(self):
         # dict is already tested via copy. Just the other cases here.